% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2k_oxy_to_tbl.R
\name{o2k_oxy_to_list}
\alias{o2k_oxy_to_list}
\title{Format Oroboros O2K Output.csv Into Lists}
\usage{
o2k_oxy_to_list(csv = "NULL")
}
\arguments{
\item{csv}{A csv file exported from the Oroboros-O2K. This should have six columns ("Time \link{s}", "Event Name", "Chamber", "Event Text", "1A: O2 flux per V \link{pmol/(s*mL)}" and "1B: O2 flux per V \link{pmol/(s*mL)})" ).}
}
\value{
A list of 3 elements; a list of raw oxygen flux measurements per chamber, opening time details, and unique events.
}
\description{
\code{o2k_oxy_to_list()} extracts data from a csv file which has been produced by DatLab and the Oroboros O2K. The output is a list with three elements. Element one is itself a list of two elements (left and right), which contain the raw data and events for the left and right chambers respectively. Element two is a tibble containing details of when the chambers were opened; the third and final element is a vector of unique events from within the run (excluding 'open', 'closed', and 'warnings'). The results of this function are designed so that they can be directly passed into \code{o2k_oxy_list_to_tbl()}.
}
\examples{
 data <- tidymito::o2k_oxy_to_list(csv = "path/to/data/file.csv")


}
\seealso{
Other read o2k files: 
\code{\link{o2k_oxy_list_to_tbl}()},
\code{\link{read_o2k_oxy_csv}()},
\code{\link{read_o2k_oxy_xlsx}()}
}
\concept{read o2k files}
