% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dams_activity.R
\name{read_dams_activity_csv}
\alias{read_dams_activity_csv}
\title{Convert Raw Drosophila Activity Monitor Outputs Into Tibble}
\usage{
read_dams_activity_csv(
  dir_path = ".",
  directory_pattern = "DAMS_b.*",
  meta_pattern = ".*meta.*",
  meta_col_1 = "monitor",
  meta_col_2 = "bloc",
  meta_date_col = 3,
  bloc_name = "DAMS_b",
  numeric_bloc = TRUE
)
}
\arguments{
\item{dir_path}{The path to the directory that holds your directories with your data of interest. By default set to the home directory.}

\item{directory_pattern}{A regex string that matches each of the directories which contain the data of interest. By default set to "DAMS_b.*".}

\item{meta_pattern}{A regex string that \emph{uniquely} matches the metadata file.}

\item{meta_col_1}{A character string with the name of the first column in the metadata file. Defaults to ".\\emph{meta.\}" (i.e will match any file with "meta" somewhere in the name.)}

\item{meta_col_2}{A character string with the name of the second column in the metadata file. Defaults to "bloc".}

\item{meta_date_col}{A single numeric value indicating the column position containing date values in the metadata file.}

\item{bloc_name}{A regex string that links the metadata bloc value and the folder name.}

\item{numeric_bloc}{A boolean value stating whether the bloc identifier values are numeric (TRUE; the default, recommended) or not (FALSE).}
}
\value{
A single tibble containing count data for every Drosophila Activity Monitor listed.
}
\description{
Drosophila Activity Monitors (a.k.a DAMs) (TriKinetics) are a valuable tool for assessing the patterns of movement in fly studies. This function aims to automate the process of extracting information from raw the datafiles produced by the instrument, removing the need for tedious (and potentially incorrect) copy pasting of data into excel files.
}
\examples{
 #
 # read_dams_activity_csv(
 #   "path/to/directory",
 #   "folder_name",
 #   ".*metafile.identifier.*",
 #   "monitor",
 #   "bloc",
 #   3,
 #   "DAMS_b)



}
