% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_o2k_oxy_xlsx.R
\name{read_o2k_oxy_xlsx}
\alias{read_o2k_oxy_xlsx}
\title{Import oxygen flux data from one or more O2k-formatted Excel Files}
\usage{
read_o2k_oxy_xlsx(input_files = "NULL", sheet = "Specific flux (bc)")
}
\arguments{
\item{input_files}{A character string (or vector of character strings) containing the path(s) to the input file(s). Every input file should follow the same format as the Oroboros template and follow the same SUIT protocol. If paths to multiple files are provided, these will be combined in the final output.}

\item{sheet}{The sheet in the supplied excel files from which the oxygen flux measurements will be retrieved. By default set to 'Specific flux (bc)'.}
}
\value{
A single tibble with oxygen flux data.
}
\description{
Import oxygen flux data from one or more O2k-formatted Excel Files
}
\seealso{
Other read o2k files: 
\code{\link{o2k_oxy_list_to_tbl}()},
\code{\link{o2k_oxy_to_list}()},
\code{\link{read_o2k_oxy_csv}()}
}
\concept{read o2k files}
